+++ src/main/java/com/cibertec/backend/qorifit/infraestructure/web/dto/response/RecipeSummaryResponse.java
@@ -11,5 +11,5 @@ public class RecipeSummaryResponse{
     private String name;
     private String countryName;
     private String imagePath;
-    private Integer estimatedCalories;
+    private Long estimatedCalories;
 }
